+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- VIMEET 2110 - Add support to broadcast live video in webinar, using an iframe
 
 ## [2.33.0] - 2020-06-22
 ### Added